feat: adicionando validação do jwt

+++ app/Http/Controllers/TaskController.php
@@ -3,10 +3,17 @@
 namespace App\Http\Controllers;
 
 use App\Models\Task;
+use App\Models\ToDoList;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class TaskController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth:api');
+    }
+
     // Criar uma nova tarefa
     public function store(Request $request)
     {
@@ -15,14 +22,23 @@ class TaskController extends Controller
             'description' => 'required|string|max:255',
         ]);
 
+        $toDoList = ToDoList::findOrFail($request->to_do_list_id);
+        
+        if ($toDoList->user_id !== Auth::id()) {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
+
         return Task::create($request->all());
     }
 
     // Atualizar uma tarefa
     public function update(Request $request, Task $task)
     {
+        if ($task->toDoList->user_id !== Auth::id()) {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
+
         $request->validate([
-            'description' => 'required|string|max:255',
             'is_completed' => 'required|boolean',
         ]);
 
@@ -33,6 +49,10 @@ class TaskController extends Controller
     // Deletar uma tarefa
     public function destroy(Task $task)
     {
+        if ($task->toDoList->user_id !== Auth::id()) {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
+
         $task->delete();
         return response()->noContent();
     }
